+++ benchmark/src/main/java/core/framework/api/util/JSONBenchmark.java
@@ -21,7 +21,7 @@ import org.openjdk.jmh.annotations.Warmup;
 @Warmup(iterations = 3)
 @Measurement(iterations = 10)
 public class JSONBenchmark {
-    private final String json = "{\"request_id\":\"d2bbf14c-aff0-402d-80a6-ed8a672ab231\",\"product\":{\"vendor_number\":\"V00090\",\"vendor_sku\":\"ELP-430205-C\",\"listing_status\":\"NOT_LIVE\",\"parent_vendor_sku\":\"ELP-430205-C\",\"name\":null,\"description\":null,\"brand_name\":null,\"category_id\":null,\"external_product_id_type\":null,\"external_product_id\":null,\"type\":null,\"unit_cost\":null,\"list_price\":null,\"map_price\":null,\"standard_price\":null,\"sale_price\":null,\"sale_start_date\":null,\"sale_end_date\":null,\"weight\":null,\"harmonized_code\":null,\"on_hold\":null,\"attributes\":{},\"eligible_country_codes\":null,\"images\":[],\"inventories\":[{\"warehouse_code\":\"V00090A\",\"quantity\":0}]},\"requested_by\":\"channel-advisor\",\"result_queue\":\"channel-advisor-service-queue\"}";
+    private static final String JSON_STRING = "{\"request_id\":\"d2bbf14c-aff0-402d-80a6-ed8a672ab231\",\"product\":{\"vendor_number\":\"V00090\",\"vendor_sku\":\"ELP-430205-C\",\"listing_status\":\"NOT_LIVE\",\"parent_vendor_sku\":\"ELP-430205-C\",\"name\":null,\"description\":null,\"brand_name\":null,\"category_id\":null,\"external_product_id_type\":null,\"external_product_id\":null,\"type\":null,\"unit_cost\":null,\"list_price\":null,\"map_price\":null,\"standard_price\":null,\"sale_price\":null,\"sale_start_date\":null,\"sale_end_date\":null,\"weight\":null,\"harmonized_code\":null,\"on_hold\":null,\"attributes\":{},\"eligible_country_codes\":null,\"images\":[],\"inventories\":[{\"warehouse_code\":\"V00090A\",\"quantity\":0}]},\"requested_by\":\"channel-advisor\",\"result_queue\":\"channel-advisor-service-queue\"}";
     private byte[] jsonBytes;
     private UpdateProductRequest object;
     private JSONReader<UpdateProductRequest> reader;
@@ -29,8 +29,8 @@ public class JSONBenchmark {
 
     @Setup
     public void setup() {
-        jsonBytes = Strings.bytes(json);
-        object = JSON.fromJSON(UpdateProductRequest.class, json);
+        jsonBytes = Strings.bytes(JSON_STRING);
+        object = JSON.fromJSON(UpdateProductRequest.class, JSON_STRING);
         reader = JSONReader.of(UpdateProductRequest.class);
         writer = JSONWriter.of(UpdateProductRequest.class);
     }
@@ -52,7 +52,7 @@ public class JSONBenchmark {
 
     @Benchmark
     public void fromJSONString() {
-        JSON.fromJSON(UpdateProductRequest.class, json);
+        JSON.fromJSON(UpdateProductRequest.class, JSON_STRING);
     }
 
     @Benchmark