+++ log-processor/src/main/java/core/log/job/CleanupOldIndexJob.java
@@ -1,22 +1,17 @@
 package core.log.job;
 
-import com.carrotsearch.hppc.cursors.ObjectObjectCursor;
 import core.framework.api.scheduler.Job;
 import core.framework.api.search.ElasticSearch;
-import org.elasticsearch.ElasticsearchException;
-import org.elasticsearch.action.admin.cluster.state.ClusterStateRequest;
-import org.elasticsearch.action.admin.cluster.state.ClusterStateResponse;
-import org.elasticsearch.action.admin.indices.close.CloseIndexRequest;
-import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
-import org.elasticsearch.client.AdminClient;
+import core.framework.api.search.ElasticSearchIndex;
+import core.framework.api.search.SearchException;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
-import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
+import java.util.List;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -34,12 +29,9 @@ public class CleanupOldIndexJob implements Job {
     public void execute() throws Exception {
         LocalDate now = LocalDate.now();
 
-        AdminClient adminClient = elasticSearch.client().admin();
-        ClusterStateResponse response = adminClient.cluster().state(new ClusterStateRequest().clear().metaData(true)).get();
-        ImmutableOpenMap<String, IndexMetaData> indices = response.getState().getMetaData().indices();
-
-        for (ObjectObjectCursor<String, IndexMetaData> cursor : indices) {
-            process(cursor.value, now);
+        List<ElasticSearchIndex> indices = elasticSearch.indices();
+        for (ElasticSearchIndex index : indices) {
+            process(index, now);
         }
     }
 
@@ -50,32 +42,29 @@ public class CleanupOldIndexJob implements Job {
         return Optional.of(LocalDate.parse(timestamp));
     }
 
-    private void process(IndexMetaData metaData, LocalDate now) {
-        String index = metaData.getIndex();
-        createdDate(index).ifPresent(date -> {
+    private void process(ElasticSearchIndex index, LocalDate now) {
+        createdDate(index.index).ifPresent(date -> {
             long days = ChronoUnit.DAYS.between(date, now);
             if (days >= 30) {        // delete log older than 30 days, close index older than 7 days
-                deleteIndex(index);
-            } else if (days >= 7 && metaData.getState() == IndexMetaData.State.OPEN) {
-                closeIndex(index);
+                deleteIndex(index.index);
+            } else if (days >= 7 && index.state == IndexMetaData.State.OPEN) {
+                closeIndex(index.index);
             }
         });
     }
 
     private void deleteIndex(String index) {
         try {
-            logger.info("delete index, index={}", index);
-            elasticSearch.client().admin().indices().delete(new DeleteIndexRequest(index)).actionGet();
-        } catch (ElasticsearchException e) {
+            elasticSearch.deleteIndex(index);
+        } catch (SearchException e) {
             logger.error(e.getMessage(), e);
         }
     }
 
     private void closeIndex(String index) {
         try {
-            logger.info("close index, index={}", index);
-            elasticSearch.client().admin().indices().close(new CloseIndexRequest(index)).actionGet();
-        } catch (ElasticsearchException e) {
+            elasticSearch.closeIndex(index);
+        } catch (SearchException e) {
             logger.error(e.getMessage(), e);
         }
     }